Consolidate Plug.Conn.Adapter calling-process tests into adapter_test.exs
Bandit.Adapter implements Plug.Conn.Adapter once, shared verbatim by
both the HTTP/1 and HTTP/2 stacks - its behaviour doesn't vary by
protocol. The "must be called by stream owner" check was tested with
duplicated (and inconsistent) coverage split across
test/bandit/http1/plug_test.exs (send_resp only) and
test/bandit/http2/plug_test.exs (read_body, send_resp, send_chunked,
chunk, send_file, inform).

Move all of it into a single test/bandit/adapter_test.exs, run
against one representative (HTTP/1) server, with matching coverage
for every relevant Plug.Conn.Adapter callback.
